Move hapi types to typings
